fix(sandbox): share container mount detection
