+++ crates/config/Cargo.toml
@@ -13,8 +13,8 @@ secrecy        = { workspace = true }
 serde          = { workspace = true }
 serde_json     = { workspace = true }
 serde_yaml     = { workspace = true }
-thiserror      = { workspace = true }
 tempfile       = { workspace = true }
+thiserror      = { workspace = true }
 toml           = { workspace = true }
 toml_edit      = { workspace = true }
 tracing        = { workspace = true }
