+++ crates/tools/src/sandbox/paths.rs
@@ -1,12 +1,12 @@
 //! Path resolution, mount detection, and home persistence directories.
 
 use std::{
-    collections::{HashMap, HashSet},
+    collections::HashMap,
     path::{Path as FsPath, PathBuf},
     sync::{Mutex, OnceLock},
 };
 
-use tracing::{debug, warn};
+use tracing::warn;
 
 use {
     super::{
@@ -19,20 +19,7 @@ use {
     crate::error::Result,
 };
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct ContainerMount {
-    pub(crate) source: PathBuf,
-    pub(crate) destination: PathBuf,
-}
-
 pub(crate) static HOST_DATA_DIR_CACHE: OnceLock<Mutex<HashMap<String, PathBuf>>> = OnceLock::new();
-#[cfg(test)]
-pub(crate) static TEST_CONTAINER_MOUNT_OVERRIDES: OnceLock<
-    Mutex<HashMap<String, Vec<ContainerMount>>>,
-> = OnceLock::new();
-#[cfg(test)]
-pub(crate) static TEST_RUNNING_CONTAINER_REFERENCES: OnceLock<Mutex<HashMap<String, Vec<String>>>> =
-    OnceLock::new();
 
 pub(crate) fn configured_host_data_dir(config: &SandboxConfig) -> Option<PathBuf> {
     let guest_data_dir = moltis_config::data_dir();
@@ -44,258 +31,6 @@ pub(crate) fn configured_host_data_dir(config: &SandboxConfig) -> Option<PathBuf
         return Some(path.clone());
     }
     Some(guest_data_dir.join(path))
-}
-
-pub(crate) fn read_trimmed_file(path: &str) -> Option<String> {
-    std::fs::read_to_string(path)
-        .ok()
-        .map(|contents| contents.trim().to_string())
-        .filter(|contents| !contents.is_empty())
-}
-
-pub(crate) fn normalize_cgroup_container_ref(segment: &str) -> Option<String> {
-    let mut value = segment.trim();
-    if value.is_empty() {
-        return None;
-    }
-    if let Some(stripped) = value.strip_suffix(".scope") {
-        value = stripped;
-    }
-    for prefix in ["docker-", "libpod-", "cri-containerd-"] {
-        if let Some(stripped) = value.strip_prefix(prefix) {
-            value = stripped;
-            break;
-        }
-    }
-    if value.len() < 12 || !value.chars().all(|ch| ch.is_ascii_hexdigit()) {
-        return None;
-    }
-    Some(value.to_string())
-}
-
-pub(crate) fn current_container_references() -> Vec<String> {
-    let mut refs = Vec::new();
-    let mut seen = HashSet::new();
-    for candidate in [
-        std::env::var("HOSTNAME").ok(),
-        read_trimmed_file("/etc/hostname"),
-    ]
-    .into_iter()
-    .flatten()
-    {
-        if seen.insert(candidate.clone()) {
-            refs.push(candidate);
-        }
-    }
-    if let Ok(cgroup) = std::fs::read_to_string("/proc/self/cgroup") {
-        for candidate in cgroup
-            .lines()
-            .flat_map(|line| line.split(['/', ':']))
-            .filter_map(normalize_cgroup_container_ref)
-        {
-            if seen.insert(candidate.clone()) {
-                refs.push(candidate);
-            }
-        }
-    }
-    refs
-}
-
-pub(crate) fn parse_container_mounts_from_inspect(stdout: &str) -> Vec<ContainerMount> {
-    let Ok(json): std::result::Result<serde_json::Value, _> = serde_json::from_str(stdout) else {
-        return Vec::new();
-    };
-    let root = json
-        .as_array()
-        .and_then(|entries| entries.first())
-        .unwrap_or(&json);
-    root.get("Mounts")
-        .and_then(serde_json::Value::as_array)
-        .into_iter()
-        .flatten()
-        .filter_map(|entry| {
-            let source = entry.get("Source")?.as_str()?.trim();
-            let destination = entry.get("Destination")?.as_str()?.trim();
-            if source.is_empty() || destination.is_empty() {
-                return None;
-            }
-            Some(ContainerMount {
-                source: PathBuf::from(source),
-                destination: PathBuf::from(destination),
-            })
-        })
-        .collect()
-}
-
-pub(crate) fn resolve_host_path_from_mounts(
-    guest_path: &FsPath,
-    mounts: &[ContainerMount],
-) -> Option<PathBuf> {
-    mounts
-        .iter()
-        .filter_map(|mount| {
-            let relative = guest_path.strip_prefix(&mount.destination).ok()?;
-            Some((
-                mount.destination.components().count(),
-                if relative.as_os_str().is_empty() {
-                    mount.source.clone()
-                } else {
-                    mount.source.join(relative)
-                },
-            ))
-        })
-        .max_by_key(|(depth, _)| *depth)
-        .map(|(_, resolved)| resolved)
-}
-
-#[cfg(test)]
-pub(crate) fn test_container_mount_override_key(cli: &str, reference: &str) -> String {
-    format!("{cli}:{reference}")
-}
-
-pub(crate) fn inspect_current_container_mounts(cli: &str, reference: &str) -> Vec<ContainerMount> {
-    #[cfg(test)]
-    {
-        let overrides = TEST_CONTAINER_MOUNT_OVERRIDES.get_or_init(|| Mutex::new(HashMap::new()));
-        let guard = overrides.lock().unwrap_or_else(|error| error.into_inner());
-        if let Some(mounts) = guard.get(&test_container_mount_override_key(cli, reference)) {
-            return mounts.clone();
-        }
-        Vec::new()
-    }
-
-    #[cfg(not(test))]
-    {
-        let output = match std::process::Command::new(cli)
-            .args(["inspect", reference])
-            .output()
-        {
-            Ok(output) if output.status.success() => output,
-            Ok(output) => {
-                let stderr = String::from_utf8_lossy(&output.stderr);
-                debug!(
-                    cli,
-                    reference,
-                    stderr = %stderr.trim(),
-                    "container inspect failed while auto-detecting host data dir"
-                );
-                return Vec::new();
-            },
-            Err(error) => {
-                debug!(
-                    cli,
-                    reference,
-                    %error,
-                    "could not inspect current container while auto-detecting host data dir"
-                );
-                return Vec::new();
-            },
-        };
-        parse_container_mounts_from_inspect(&String::from_utf8_lossy(&output.stdout))
-    }
-}
-
-pub(crate) fn running_container_references(cli: &str) -> Vec<String> {
-    #[cfg(test)]
-    {
-        return TEST_RUNNING_CONTAINER_REFERENCES
-            .get_or_init(|| Mutex::new(HashMap::new()))
-            .lock()
-            .unwrap_or_else(|error| error.into_inner())
-            .get(cli)
-            .cloned()
-            .unwrap_or_default();
-    }
-
-    #[cfg(not(test))]
-    {
-        let output = match std::process::Command::new(cli)
-            .args(["ps", "-q", "--no-trunc"])
-            .output()
-        {
-            Ok(output) if output.status.success() => output,
-            Ok(output) => {
-                let stderr = String::from_utf8_lossy(&output.stderr);
-                debug!(
-                    cli,
-                    stderr = %stderr.trim(),
-                    "container list failed while auto-detecting host data dir"
-                );
-                return Vec::new();
-            },
-            Err(error) => {
-                debug!(
-                    cli,
-                    %error,
-                    "could not list containers while auto-detecting host data dir"
-                );
-                return Vec::new();
-            },
-        };
-        String::from_utf8_lossy(&output.stdout)
-            .lines()
-            .map(str::trim)
-            .filter(|line| !line.is_empty())
-            .map(str::to_string)
-            .collect()
-    }
-}
-
-pub(crate) fn detect_host_data_dir_from_references(
-    cli: &str,
-    guest_data_dir: &FsPath,
-    references: &[String],
-) -> Option<PathBuf> {
-    let mut detected: Option<PathBuf> = None;
-    for reference in references {
-        let mounts = inspect_current_container_mounts(cli, reference);
-        if mounts.is_empty() {
-            continue;
-        }
-        let Some(resolved) = resolve_host_path_from_mounts(guest_data_dir, &mounts) else {
-            continue;
-        };
-        if let Some(existing) = &detected
-            && existing != &resolved
-        {
-            debug!(
-                cli,
-                guest_path = %guest_data_dir.display(),
-                first_host_path = %existing.display(),
-                other_host_path = %resolved.display(),
-                "ambiguous host data dir from container mounts"
-            );
-            return None;
-        }
-        detected = Some(resolved);
-    }
-    detected
-}
-
-pub(crate) fn detect_host_data_dir_with_references(
-    cli: &str,
-    guest_data_dir: &FsPath,
-    references: &[String],
-) -> Option<PathBuf> {
-    if let Some(resolved) = detect_host_data_dir_from_references(cli, guest_data_dir, references) {
-        debug!(
-            cli,
-            guest_path = %guest_data_dir.display(),
-            host_path = %resolved.display(),
-            "auto-detected host data dir from current container mounts"
-        );
-        return Some(resolved);
-    }
-
-    let running_references = running_container_references(cli);
-    let resolved = detect_host_data_dir_from_references(cli, guest_data_dir, &running_references)?;
-    debug!(
-        cli,
-        guest_path = %guest_data_dir.display(),
-        host_path = %resolved.display(),
-        "auto-detected host data dir by scanning running container mounts"
-    );
-    Some(resolved)
 }
 
 pub(crate) fn host_data_dir_cache() -> &'static Mutex<HashMap<String, PathBuf>> {
@@ -313,8 +48,11 @@ pub(crate) fn detect_host_data_dir(cli: &str, guest_data_dir: &FsPath) -> Option
         }
     }
 
-    let detected =
-        detect_host_data_dir_with_references(cli, guest_data_dir, &current_container_references());
+    let detected = moltis_config::container_mounts::detect_host_data_dir_with_references(
+        cli,
+        guest_data_dir,
+        &moltis_config::container_mounts::current_container_references(),
+    );
 
     if let Some(path) = detected.clone() {
         let mut guard = host_data_dir_cache()
